backend/requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
flask-cors==3.0.10
pymongo==4.6.2
pytest==8.3.5
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
requests==2.31.0
Werkzeug==2.0.3
beautifulsoup4==4.12.2
Pillow==10.2.0 
